+++ src/localisation/german.ts
@@ -582,7 +582,12 @@ const GERMAN: Translations = {
   [Phrase.LatestCombatLogFileOverAMonthOld]: 'über einen Monat alt',
   [Phrase.LatestCombatLogFileDayOld]: '1 Tag alt',
   [Phrase.LatestCombatLogFileDaysOld]: 'Tage alt',
-  [Phrase.LatestCombatLogFileLessThanADayOld]: 'weniger als einen Tag alt'
+  [Phrase.LatestCombatLogFileLessThanADayOld]: 'weniger als einen Tag alt',
+  [Phrase.HotKeySettingsLabel]: 'Tastenkürzel',
+  [Phrase.ForceStopHotKeyLabel]: 'Stopp erzwingen Hotkey',
+  [Phrase.ForceStopHotKeyDescription]: 'Setze einen Hotkey, um die laufende Aufnahme zwangsweise zu beenden.',
+  [Phrase.HotKeyConflictText]:
+    'Diese Tastenkombination wird bereits von einem anderen Hotkey verwendet.',
 };
 
 export default GERMAN;